+++ Sources/BraintreeApplePay/Public/BraintreeApplePay/BTApplePayCardNonce.h
@@ -17,9 +17,14 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, readonly, strong) BTBinData *binData;
 
 /**
+ The last two digits of the card associated with this nonce, if available.
+ */
+@property (nonatomic, nullable, readonly, copy) NSString *dpanLastTwo;
+
+/**
  Used to initialize a `BTApplePayCardNonce` with parameters.
  */
-- (nullable instancetype)initWithNonce:(NSString *)nonce type:(NSString *)type json:(BTJSON *)json;
+- (nullable instancetype)initWithJSON:(BTJSON *)json;
 
 @end